move dynamic test creation to ./utils.js
